Fixing Xcode 5.1.1 warnings.

+++ iPhone/iTurms/IAMAppDelegate.h
@@ -21,7 +21,7 @@
 
 // helpers for corelocation.
 @property (nonatomic, assign) BOOL isLocationDenied;
-@property (nonatomic, assign) int nLocationUseDenies;
+@property (nonatomic, assign) NSInteger nLocationUseDenies;
 @property (nonatomic, strong) CLLocationManager *locationManager;
 @property (nonatomic) NSString *locationString;
